requirements.txt
amqp==5.1.1
asgiref==3.5.1
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.6
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==3.2.13
django-celery-beat==2.2.1
django-environ==0.8.1
django-js-asset==2.0.0
django-mptt==0.13.4
django-seed==0.3.1
django-timezone-field==4.2.3
djangorestframework==3.13.1
Faker==13.11.0
kombu==5.2.4
packaging==21.3
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.1
redis==4.3.1
six==1.16.0
sqlparse==0.4.2
toposort==1.7
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
